Refactor) roomSocket enterUpdate
